fix dashboard structure
