+++ app/config/api/.env.example
@@ -1,0 +1,3 @@
+# https://coinmarketcap.com/api/pricing/
+API_KEY=
+API_URL=
